refactor(dgw): re-organize netscan-related modules
