+++ devolutions-gateway/src/api/mod.rs
@@ -7,7 +7,7 @@ pub mod jmux;
 pub mod jrec;
 pub mod jrl;
 pub mod kdc_proxy;
-pub mod network_scan;
+pub mod net;
 pub mod rdp;
 pub mod session;
 pub mod sessions;
@@ -27,7 +27,7 @@ pub fn make_router<S>(state: crate::DgwState) -> axum::Router<S> {
         .route("/jet/rdp", axum::routing::get(rdp::handler))
         .nest("/jet/fwd", fwd::make_router(state.clone()))
         .nest("/jet/webapp", webapp::make_router(state.clone()))
-        .nest("/jet/net", network_scan::make_router(state.clone()));
+        .nest("/jet/net", net::make_router(state.clone()));
 
     if state.conf_handle.get_conf().webapp_is_enabled() {
         router = router.route(
