--- devolutions-gateway/src/api/network_scan/mod.rs
@@ -1,12 +1,0 @@
-use crate::DgwState;
-use axum::Router;
-
-pub mod ipconfig;
-pub mod scan;
-
-pub fn make_router<S>(state: DgwState) -> Router<S> {
-    Router::new()
-        .route("/scan", axum::routing::get(scan::handler))
-        .route("/config", axum::routing::get(ipconfig::handler))
-        .with_state(state)
-}
